hangul_combination_set_data()에서 size_t overflow 문제 수정
대부분의 경우 ULONG_MAX는 SIZE_MAX와 같은 값을 가지겠지만,
size_t 타입에 맞는 최대값을 사용하는 편이 안전할 것이다.

https://github.com/choehwanjin/libhangul/issues/17

+++ hangul/hangulkeyboard.c
@@ -275,7 +275,7 @@ hangul_combination_set_data(HangulCombination* combination,
     if (combination == NULL)
 	return false;
 
-    if (n == 0 || n > ULONG_MAX / sizeof(HangulCombinationItem))
+    if (n == 0 || n > SIZE_MAX / sizeof(HangulCombinationItem))
 	return false;
 
     combination->table = malloc(sizeof(HangulCombinationItem) * n);
